Separate each class in SettingsMenu.cs to its own classes for readability
